requirements.txt
numpy
faiss-cpu
scikit-learn
streamlit
unstructured[pdf]
sentence-transformers
transformers
google-generativeai
pytesseract